Move number literal parsing to own submodule
